pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sr-training"
version = "0.1.0"
description = "Optimized Super Resolution Training"
authors = [{name = "SR Training", email = "training@example.com"}]
requires-python = ">=3.9"
dependencies = [
    # Core PyTorch packages
    "torch>=1.12.0",
    "torchvision>=0.13.0", 
    "torchaudio>=0.12.0",
    
    # Data processing
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "pillow>=8.3.0",
    "opencv-python>=4.5.0",
    "lmdb>=1.2.0",
    
    # Progress and visualization
    "tqdm>=4.62.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    
    # Logging and monitoring
    "tensorboardx>=2.4.0",
    "wandb>=0.12.0",
    
    # Utilities
    "scipy>=1.7.0",
    "scikit-image>=0.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=22.0",
    "isort>=5.0",
    "flake8>=4.0",
]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100
